requirements.txt
openai>=1.0.0
tiktoken 
flask
flask-socketio==5.3.0
GitPython
python-dotenv
requests
beautifulsoup4
chromadb
gunicorn